Implement all git operations in terms of `Storage`
Making `Repo` just a thin, namespace-aware wrapper
