+++ librad/src/git/repo.rs
@@ -15,33 +15,19 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::collections::HashSet;
 
-use serde::{de::DeserializeOwned, Serialize};
 use thiserror::Error;
 
 use crate::{
     git::{
-        ext::{Git2ErrorExt, Oid, References},
-        refs::{self, Refs},
-        storage::{self, Storage, WithBlob},
-        types::{Namespace, Reference, Refspec},
-        url::GitUrlRef,
+        refs::Refs,
+        storage::{self, Storage},
+        types::Namespace,
     },
-    hash::Hash,
-    internal::{
-        borrow::{TryCow, TryToOwned},
-        canonical::{Cjson, CjsonError},
-    },
-    meta::entity::{
-        self,
-        data::{EntityBuilder, EntityData},
-        Draft,
-        Entity,
-        Signatory,
-    },
+    internal::borrow::{TryCow, TryToOwned},
     peer::PeerId,
-    uri::{self, RadUrl, RadUrn},
+    uri::RadUrn,
 };
 use radicle_surf::vcs::git as surf;
 
@@ -49,243 +35,51 @@ pub use storage::Tracked;
 
 #[derive(Debug, Error)]
 pub enum Error {
-    #[error("Unknown repo {0}")]
-    NoSuchUrn(RadUrn),
-
-    #[error(
-        "Identity root hash doesn't match resolved URL. Expected {expected}, actual: {actual}"
-    )]
-    RootHashMismatch { expected: Hash, actual: Hash },
-
-    #[error("Metadata is not signed")]
-    UnsignedMetadata,
-
-    #[error(transparent)]
-    Urn(#[from] uri::rad_urn::ParseError),
-
-    #[error(transparent)]
-    Entity(#[from] entity::Error),
-
-    #[error(transparent)]
-    Refsig(#[from] refs::signed::Error),
-
-    #[error(transparent)]
-    Cjson(#[from] CjsonError),
-
     #[error(transparent)]
     Storage(#[from] storage::Error),
 
     #[error(transparent)]
     Git(#[from] git2::Error),
-
-    #[error(transparent)]
-    Surf(#[from] surf::error::Error),
 }
 
+/// A logical repository.
+///
+/// This is just a (thin) wrapper around [`Storage`] so the [`RadUrn`] context
+/// doesn't need to be passed around.
 pub struct Repo<'a> {
     pub urn: RadUrn,
-    storage: TryCow<'a, Storage>,
+    pub(super) storage: TryCow<'a, Storage>,
 }
 
 impl<'a> Repo<'a> {
-    pub fn create<T>(storage: &'a Storage, meta: &Entity<T, Draft>) -> Result<Self, Error>
-    where
-        T: Serialize + DeserializeOwned + Clone + Default,
-        EntityData<T>: EntityBuilder,
-    {
-        let span = tracing::info_span!("Repo::create");
-        let _guard = span.enter();
-
-        // FIXME: properly verify meta
-
-        if meta.signatures().is_empty() {
-            return Err(Error::UnsignedMetadata);
-        }
-
-        // FIXME: certifier identities must exist, or be supplied
-
-        let this = Self {
-            urn: RadUrn::new(
-                meta.root_hash().to_owned(),
-                uri::Protocol::Git,
-                uri::Path::empty(),
-            ),
-            storage: storage.into(),
-        };
-        this.commit_initial_meta(meta)?;
-        this.track_signers(&meta)?;
-        this.update_refs()?;
-
-        Ok(this)
+    pub fn namespace(&self) -> Namespace {
+        self.urn.id.clone()
     }
 
-    pub fn open(storage: &'a Storage, urn: RadUrn) -> Result<Self, Error> {
-        {
-            let id_ref = Reference::rad_id(urn.id.clone());
-            if !storage.has_ref(&id_ref)? {
-                return Err(Error::NoSuchUrn(urn));
-            }
-        }
-
-        Ok(Self {
-            urn: RadUrn {
-                path: uri::Path::empty(),
-                ..urn
-            },
-            storage: storage.into(),
-        })
-    }
-
-    pub fn clone<T>(storage: &'a Storage, url: RadUrl) -> Result<Self, Error>
-    where
-        T: Serialize + DeserializeOwned + Clone + Default,
-        EntityData<T>: EntityBuilder,
-    {
-        let span = tracing::info_span!("Repo::clone", repo.url = %url);
-        let _guard = span.enter();
-
-        let local_peer_id = PeerId::from(&storage.key);
-        let this = Self {
-            urn: RadUrn {
-                path: uri::Path::empty(),
-                ..url.urn.clone()
-            },
-            storage: storage.into(),
-        };
-
-        // Fetch the identity first
-        let git_url = GitUrlRef::from_rad_url_ref(url.as_ref(), &local_peer_id);
-        let meta = this.fetch_id(git_url)?;
-
-        // TODO: properly verify meta
-
-        if meta.signatures().is_empty() {
-            return Err(Error::UnsignedMetadata);
-        }
-
-        if meta.root_hash() != &url.urn.id {
-            return Err(Error::RootHashMismatch {
-                expected: url.urn.id.to_owned(),
-                actual: meta.root_hash().to_owned(),
-            });
-        }
-
-        this.track_signers(&meta)?;
-        this.update_refs()?;
-        this.fetch(&url.authority)?;
-
-        Ok(this)
-    }
-
-    pub fn browser(&'_ self) -> Result<surf::Browser<'_>, Error> {
-        Ok(surf::Browser::new(&self.storage.backend)?)
-    }
-
+    /// Fetch new refs and objects for this repo from [`PeerId`]
     pub fn fetch(&self, from: &PeerId) -> Result<(), Error> {
-        let span = tracing::info_span!("Repo::fetch", repo.fetch.from = %from);
-        let _guard = span.enter();
-
-        let namespace = &self.urn.id;
-
-        let mut remote = {
-            let local_peer = PeerId::from(&self.storage.key);
-            let url = GitUrlRef::from_rad_url_ref(self.urn.as_rad_url_ref(from), &local_peer);
-            self.storage.remote_anonymous(&url.to_string())
-        }?;
-        remote.connect(git2::Direction::Fetch)?;
-
-        let rad_refs = self.rad_refs()?;
-        let tracked_trans = rad_refs.remotes.flatten().collect::<HashSet<&PeerId>>();
-
-        // Fetch rad/refs of all known remotes
-        {
-            let refspecs =
-                Refspec::rad_refs(namespace.clone(), from, tracked_trans.iter().cloned())
-                    .map(|spec| spec.to_string())
-                    .collect::<Vec<String>>();
-            tracing::debug!(refspecs = ?refspecs, "Fetching rad/refs");
-            remote.fetch(&refspecs, Some(&mut self.fetch_options()), None)?;
-        }
-
-        // Read the signed refs of all known remotes, and compare their `heads`
-        // against the advertised refs. If signed and advertised branch head
-        // match, non-fast-forwards are permitted. Otherwise, the branch is
-        // skipped.
-        {
-            let remote_heads: HashMap<&str, git2::Oid> = remote
-                .list()?
-                .iter()
-                .map(|rhead| (rhead.name(), rhead.oid()))
-                .collect();
-
-            let refspecs = Refspec::fetch_heads(
-                namespace.clone(),
-                remote_heads,
-                tracked_trans.iter().cloned(),
-                from,
-                |peer| self.rad_refs_of(peer),
-                |peer| self.certifiers_of(peer),
-            )?
-            .map(|spec| spec.to_string())
-            .collect::<Vec<String>>();
-
-            tracing::debug!(refspecs = ?refspecs, "Fetching refs/heads");
-            remote.fetch(&refspecs, Some(&mut self.fetch_options()), None)?;
-        }
-
-        // At this point, the transitive tracking graph may have changed. Let's
-        // update the refs, but don't recurse here for now (we could, if
-        // we reload `self.refs()` and compare to the value we had
-        // before fetching).
-        self.update_refs()
+        self.storage
+            .fetch_repo(&self.urn, from)
+            .map_err(Error::from)
     }
 
+    /// Obtain a read-only view of this repo
+    pub fn browser(&'_ self) -> Result<surf::Browser<'_>, Error> {
+        self.storage.browser(&self.urn).map_err(Error::from)
+    }
+
+    /// Track [`PeerId`]s view of this repo
+    ///
+    /// Equivalent to `git remote add`.
     pub fn track(&self, peer: &PeerId) -> Result<(), Error> {
-        self.storage.track(&self.urn, peer)?;
-        Ok(())
+        self.storage.track(&self.urn, peer).map_err(Error::from)
     }
 
-    // FIXME: decide if we want to require verified entities
-    fn track_signers<T>(&self, meta: &Entity<T, Draft>) -> Result<(), Error>
-    where
-        T: Serialize + DeserializeOwned + Clone + Default,
-    {
-        let span = tracing::debug_span!("Repo::track_signers", meta.urn = %meta.urn());
-        let _guard = span.enter();
-
-        meta.signatures()
-            .iter()
-            .map(|(pk, sig)| {
-                let peer_id = PeerId::from(pk.clone());
-                match &sig.by {
-                    Signatory::User(urn) => (peer_id, Some(urn)),
-                    Signatory::OwnedKey => (peer_id, None),
-                }
-            })
-            .try_for_each(|(peer, urn)| {
-                tracing::debug!(
-                    tracked.peer = %peer,
-                    tracked.urn =
-                        %urn.map(|urn| urn.to_string()).unwrap_or_else(|| "None".to_owned()),
-                    "Tracking signer of {}",
-                    meta.urn()
-                );
-
-                // Track the signer's version of this repo (if any)
-                self.track(&peer)?;
-                // Track the signer's version of the identity she used for
-                // signing (if any)
-                if let Some(urn) = urn {
-                    self.storage.track(urn, &peer)?;
-                }
-
-                Ok(())
-            })
-    }
-
+    /// Stop tracking [`PeerId`]s view of this repo
+    ///
+    /// Equivalent to `git remote rm`.
     pub fn untrack(&self, peer: &PeerId) -> Result<(), Error> {
-        self.storage.untrack(&self.urn, peer)?;
-        Ok(())
+        self.storage.untrack(&self.urn, peer).map_err(Error::from)
     }
 
     /// Retrieve all _directly_ tracked peers
@@ -293,83 +87,32 @@ impl<'a> Repo<'a> {
     /// To retrieve the transitively tracked peers, use [`rad_refs`] and inspect
     /// the `remotes`.
     pub fn tracked(&self) -> Result<Tracked, Error> {
-        let tracked = self.storage.tracked(&self.urn)?;
-        Ok(tracked)
+        self.storage.tracked(&self.urn).map_err(Error::from)
     }
 
-    /// Read the current [`Refs`] from the repo state
+    /// Retrieve all directly _as well_ as transitively tracked peers
     pub fn rad_refs(&self) -> Result<Refs, Error> {
-        let span = tracing::debug_span!("Repo::rad_refs", urn = %self.urn);
-        let _guard = span.enter();
-
-        // Collect refs/heads (our branches) at their current state
-        let heads = self.references_glob(Some("refs/heads/*"))?;
-        let heads: BTreeMap<String, Oid> = heads.map(|(name, oid)| (name, Oid(oid))).collect();
-
-        tracing::debug!(heads = ?heads);
-
-        // Get 1st degree tracked peers from the remotes configured in .git/config
-        let tracked = self.tracked()?;
-        let mut remotes: HashMap<PeerId, HashMap<PeerId, HashSet<PeerId>>> =
-            tracked.map(|peer| (peer, HashMap::new())).collect();
-
-        tracing::debug!(remotes.bare = ?remotes);
-
-        // For each of the 1st degree tracked peers, lookup their rad/refs (if any),
-        // verify the signature, and add their [`Remotes`] to ours (minus the 3rd
-        // degree)
-        for (peer, tracked) in remotes.iter_mut() {
-            match self.rad_refs_of(peer.clone()) {
-                Ok(refs) => *tracked = refs.remotes.cutoff(),
-                Err(Error::Storage(storage::Error::NoSuchBranch(_)))
-                | Err(Error::Storage(storage::Error::NoSuchBlob(_))) => {},
-                Err(e) => return Err(e),
-            }
-        }
-
-        tracing::debug!(remotes.verified = ?remotes);
-
-        Ok(Refs {
-            heads,
-            remotes: remotes.into(),
-        })
+        self.storage.rad_refs(&self.urn).map_err(Error::from)
     }
 
-    /// The set of all certifiers of this repo's identity, transitively
+    /// Retrieve the certifier URNs of this repo's identity
     pub fn certifiers(&self) -> Result<HashSet<RadUrn>, Error> {
-        let mut refs = References::from_globs(
-            &self.storage,
-            &[
-                format!("refs/namespaces/{}/refs/rad/ids/*", &self.urn.id),
-                format!("refs/namespaces/{}/refs/remotes/**/rad/ids/*", &self.urn.id),
-            ],
-        )?;
-        let refnames = refs.names();
-        Ok(urns_from_refs(refnames).collect())
+        self.storage.certifiers(&self.urn).map_err(Error::from)
     }
 
-    pub fn namespace(&self) -> Namespace {
-        self.urn.id.clone()
+    /// Check if the given [`git2::Oid`] exists within the context of this repo
+    pub fn has_commit(&self, oid: git2::Oid) -> Result<bool, Error> {
+        self.storage.has_commit(&self.urn, oid).map_err(Error::from)
     }
+
+    // TODO: find a better way to expose low-level git operations
 
     pub fn index(&self) -> Result<git2::Index, Error> {
-        let idx = self.storage.index()?;
-        Ok(idx)
+        self.storage.index().map_err(Error::from)
     }
 
     pub fn find_tree(&self, oid: git2::Oid) -> Result<git2::Tree, Error> {
-        let tree = self.storage.find_tree(oid)?;
-        Ok(tree)
-    }
-
-    pub fn blob(&self, data: &[u8]) -> Result<git2::Oid, Error> {
-        let oid = self.storage.blob(data)?;
-        Ok(oid)
-    }
-
-    pub fn find_blob(&self, oid: git2::Oid) -> Result<git2::Blob, Error> {
-        let blob = self.storage.find_blob(oid)?;
-        Ok(blob)
+        self.storage.find_tree(oid).map_err(Error::from)
     }
 
     pub fn commit(
@@ -379,221 +122,9 @@ impl<'a> Repo<'a> {
         tree: &git2::Tree,
         parents: &[&git2::Commit],
     ) -> Result<git2::Oid, Error> {
-        let author = self.storage.signature()?;
-        let head = Reference::head(self.namespace(), None, branch);
-        let oid = self.storage.commit(
-            Some(&head.to_string()),
-            &author,
-            &author,
-            msg,
-            tree,
-            parents,
-        )?;
-
-        self.update_refs()?;
-
-        Ok(oid)
-    }
-
-    pub fn find_commit(&self, oid: git2::Oid) -> Result<git2::Commit, Error> {
-        let commit = self.storage.find_commit(oid)?;
-        Ok(commit)
-    }
-
-    pub fn references_glob(
-        &'a self,
-        globs: impl IntoIterator<Item = impl AsRef<str>>,
-    ) -> Result<impl Iterator<Item = (String, git2::Oid)> + 'a, Error> {
-        let namespace_prefix = format!("refs/namespaces/{}/", &self.urn.id);
-
-        let refs = References::from_globs(
-            &self.storage,
-            globs
-                .into_iter()
-                .map(|glob| format!("{}{}", namespace_prefix, glob.as_ref())),
-        )?;
-
-        Ok(refs.peeled().filter_map(move |(name, target)| {
-            name.strip_prefix(&namespace_prefix)
-                .map(|name| (name.to_owned(), target))
-        }))
-    }
-
-    fn commit_initial_meta<T>(&self, meta: &Entity<T, Draft>) -> Result<git2::Oid, Error>
-    where
-        T: Serialize + DeserializeOwned + Clone + Default,
-        EntityData<T>: EntityBuilder,
-    {
-        let canonical_data = Cjson(meta).canonical_form()?;
-        let blob = self.storage.blob(&canonical_data)?;
-        let tree = {
-            let mut builder = self.storage.treebuilder(None)?;
-            builder.insert("id", blob, 0o100_644)?;
-            let oid = builder.write()?;
-            self.storage.find_tree(oid)
-        }?;
-        let author = self.storage.signature()?;
-
-        let branch_name = Reference::rad_id(self.namespace());
-
-        let oid = self.storage.commit(
-            Some(&branch_name.to_string()),
-            &author,
-            &author,
-            &format!("Initialised with identity {}", meta.root_hash()),
-            &tree,
-            &[],
-        )?;
-
-        tracing::debug!(
-            repo.urn = %self.urn,
-            repo.id.branch = %branch_name,
-            repo.id.oid = %oid,
-            "Initial metadata committed"
-        );
-
-        Ok(oid)
-    }
-
-    // FIXME: decide if we want to require verified entities
-    fn fetch_id<T>(&self, url: GitUrlRef) -> Result<Entity<T, Draft>, Error>
-    where
-        T: Serialize + DeserializeOwned + Clone + Default,
-        EntityData<T>: EntityBuilder,
-    {
-        tracing::debug!("Fetching id of {}", url);
-
-        let id_branch = Reference::rad_id(self.namespace());
-        let certifiers_glob = Reference::rad_ids_glob(self.namespace());
-
-        // Map rad/id to rad/id (not remotes/X/rad/id) -- we need an owned
-        // id, and the remote one is supposed to be valid regardless of the
-        // peer we're cloning from. A resolver may later decide whether it's
-        // up-to-date.
-        let refspecs = [
-            Refspec {
-                remote: id_branch.clone(),
-                local: id_branch.clone(),
-                force: false,
-            },
-            Refspec {
-                remote: certifiers_glob.clone(),
-                local: certifiers_glob,
-                force: false,
-            },
-        ]
-        .iter()
-        .map(|spec| spec.to_string())
-        .collect::<Vec<String>>();
-
-        {
-            tracing::trace!(repo.clone.refspecs = ?refspecs);
-            let mut remote = self.storage.remote_anonymous(&url.to_string())?;
-            remote.fetch(&refspecs, Some(&mut self.fetch_options()), None)?;
-        }
-
-        let entity: Entity<T, Draft> = {
-            let blob = WithBlob::Init {
-                reference: &id_branch,
-                file_name: "id",
-            }
-            .get(&self.storage)?;
-            Entity::<T, Draft>::from_json_slice(blob.content())
-        }?;
-
-        Ok(entity)
-    }
-
-    fn rad_refs_of(&self, peer: PeerId) -> Result<Refs, Error> {
-        let signed = {
-            let refs = Reference::rad_refs(self.namespace(), peer.clone());
-            let blob = WithBlob::Tip {
-                reference: &refs,
-                file_name: "refs",
-            }
-            .get(&self.storage)?;
-            refs::Signed::from_json(blob.content(), &peer)
-        }?;
-
-        Ok(Refs::from(signed))
-    }
-
-    fn update_refs(&self) -> Result<(), Error> {
-        let span = tracing::debug_span!("Repo::update_refs");
-        let _guard = span.enter();
-
-        let refsig_canonical = self
-            .rad_refs()?
-            .sign(&self.storage.key)
-            .and_then(|signed| Cjson(signed).canonical_form())?;
-
-        let rad_refs_ref = Reference::rad_refs(self.namespace(), None).to_string();
-
-        let parent: Option<git2::Commit> = self
-            .storage
-            .find_reference(&rad_refs_ref)
-            .and_then(|refs| refs.peel_to_commit().map(Some))
-            .map_not_found::<Error, _>(|| Ok(None))?;
-        let tree = {
-            let blob = self.storage.blob(&refsig_canonical)?;
-            let mut builder = self.storage.treebuilder(None)?;
-
-            builder.insert("refs", blob, 0o100_644)?;
-            let oid = builder.write()?;
-
-            self.storage.find_tree(oid)
-        }?;
-
-        // Don't create a new commit if it would be the same tree as the parent
-        if let Some(ref parent) = parent {
-            if parent.tree()?.id() == tree.id() {
-                return Ok(());
-            }
-        }
-
-        let author = self.storage.signature()?;
-        self.storage.commit(
-            Some(&rad_refs_ref),
-            &author,
-            &author,
-            "",
-            &tree,
-            &parent.iter().collect::<Vec<&git2::Commit>>(),
-        )?;
-
-        Ok(())
-    }
-
-    fn certifiers_of(&self, peer: &PeerId) -> Result<HashSet<RadUrn>, Error> {
-        let mut refs = References::from_globs(
-            &self.storage,
-            &[format!(
-                "refs/namespaces/{}/refs/remotes/{}/rad/ids/*",
-                &self.urn.id, peer
-            )],
-        )?;
-        let refnames = refs.names();
-        Ok(urns_from_refs(refnames).collect())
-    }
-
-    fn fetch_options(&self) -> git2::FetchOptions<'a> {
-        let mut cbs = git2::RemoteCallbacks::new();
-        cbs.sideband_progress(|prog| {
-            tracing::trace!("{}", unsafe { std::str::from_utf8_unchecked(prog) });
-            true
-        })
-        .update_tips(|name, old, new| {
-            tracing::debug!("{}: {} -> {}", name, old, new);
-            true
-        });
-
-        let mut fos = git2::FetchOptions::new();
-        fos.prune(git2::FetchPrune::Off)
-            .update_fetchhead(true)
-            .download_tags(git2::AutotagOption::None)
-            .remote_callbacks(cbs);
-
-        fos
+        self.storage
+            .commit(&self.urn, branch, msg, tree, parents)
+            .map_err(Error::from)
     }
 }
 
@@ -606,15 +137,4 @@ impl TryToOwned for Repo<'_> {
         let urn = self.urn.clone();
         Ok(Self { storage, urn })
     }
-}
-
-fn urns_from_refs<'a, E>(
-    refs: impl Iterator<Item = Result<&'a str, E>> + 'a,
-) -> impl Iterator<Item = RadUrn> + 'a {
-    refs.filter_map(|refname| {
-        refname
-            .ok()
-            .and_then(|name| name.split('/').next_back())
-            .and_then(|urn| urn.parse().ok())
-    })
 }
